requirements.txt
Django==3.2.6
celery==5.2.3
redis==3.5.3
djoser==2.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.3
drf-yasg==1.21.4
django-rest-swagger==2.2.0